+++ content.js
@@ -1,3 +1,6 @@
+const chrome_v2 = chrome.extension;
+const chrome_v3 = chrome.runtime; /// manifest v3
+
 const MAIN_URI = "https://v-search.nid.naver.com/reservation";
 const PROGRESS_PATH = "/progress";
 const RELOAD_INTERVAL_MILLISECONDS = 2000;
@@ -92,21 +95,21 @@ const macro = (data) => {
 
     if (location.href.includes("error")) {
         console.log("예약 시도했지만 결과페이지는 에러, 질병 관리청 응답 지연");
-        chrome.runtime.sendMessage({type: "tryButErrorTicketing", name: sCurrentName});
+        chrome_v2.sendMessage({type: "tryButErrorTicketing", name: sCurrentName});
         macroStop("try_error");
         return;
     }
 
     if (location.href.includes("failure")) {
         console.log("예약 실패쓰");
-        chrome.runtime.sendMessage({type: "failTicketing", name: sCurrentName});
+        chrome_v2.sendMessage({type: "failTicketing", name: sCurrentName});
         macroStop("fail");
         return;
     }
 
     if (location.href.includes("success")) {
         console.log("예약 성공쓰");
-        chrome.runtime.sendMessage({type: "successTicketing", name: sCurrentName});
+        chrome_v2.sendMessage({type: "successTicketing", name: sCurrentName});
         macroStop("success");
         return;
     }
@@ -134,7 +137,7 @@ const macro = (data) => {
 
         if (document.querySelector(".error_area")) {
             console.log("error detected. try reload.");
-            chrome.runtime.sendMessage({type: "errorWhileTicketing", name: sCurrentName});
+            chrome_v2.sendMessage({type: "errorWhileTicketing", name: sCurrentName});
         }
 
         _reload(data);
@@ -173,11 +176,11 @@ const macro = (data) => {
                 if (!isTest) {
                     confirmButton.click();
                     document.location = url;
-                    chrome.runtime.sendMessage({type: "tryTicketing", name: sCurrentName});
+                    chrome_v2.sendMessage({type: "tryTicketing", name: sCurrentName});
                 }
                 if (isTest) {
                     console.log("will move to :" + url);
-                    chrome.runtime.sendMessage({type: "testTicketing", url: url, name: sCurrentName});
+                    chrome_v2.sendMessage({type: "testTicketing", url: url, name: sCurrentName});
                     setTimeout(() => {
                         confirmButton.click();
                         document.location = url;